user access token moved to ram; socket auth guard added;

+++ backend/src/users/controllers/relations.controller.ts
@@ -7,6 +7,7 @@ import {
   Param,
   Post,
   Req,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { Controller } from '@nestjs/common/decorators/core/controller.decorator';
 import { ApiBearerAuth } from '@nestjs/swagger';
@@ -16,11 +17,11 @@ import {
   ApiResponse,
   ApiTags,
 } from '@nestjs/swagger/dist';
+import { JwtAuth } from '../../auth/guards/jwt-auth.guard';
 import { Profile } from '../../profiles/entities/profile.entity';
 import { profileToProfileResponse } from '../../profiles/utils/entity-payload-converter';
 import { Pair } from '../../utils/pair';
-import { JwtAuth } from '../../auth/guards/jwt-auth.guard';
-import { FriendDto as FriendDto } from '../dto/payload/add-friend-payload.dto';
+import { FriendDto } from '../dto/payload/add-friend-payload.dto';
 import { FriendsResponse } from '../dto/response/friends-response.dto';
 import {
   FriendsStatus,
@@ -47,7 +48,8 @@ export class RelationsController {
   })
   @HttpCode(HttpStatus.OK)
   @Post('/friend')
-  async addFriend(@Req() req: any, @Body() friendDto: FriendDto) {
+  async addFriend(@Req() req: Express.Request, @Body() friendDto: FriendDto) {
+    if (req.user === undefined) throw new UnauthorizedException();
     await this.relationsService.addFriend(
       req.user.userId,
       friendDto.friend_username,
@@ -63,7 +65,11 @@ export class RelationsController {
   })
   @HttpCode(HttpStatus.OK)
   @Post('/unfriend')
-  async removeFriend(@Req() req: any, @Body() friendDto: FriendDto) {
+  async removeFriend(
+    @Req() req: Express.Request,
+    @Body() friendDto: FriendDto,
+  ) {
+    if (req.user === undefined) throw new UnauthorizedException();
     await this.relationsService.removeFriend(
       req.user.userId,
       friendDto.friend_username,
@@ -76,11 +82,14 @@ export class RelationsController {
   })
   @ApiOperation({ summary: 'Get Friends List of current user' })
   @Get('/friends')
-  async getFriends(@Req() req: any): Promise<FriendsResponse[]> {
+  async getFriends(@Req() req: Express.Request): Promise<FriendsResponse[]> {
+    if (req.user === undefined) throw new UnauthorizedException();
     const pairs: Pair<UserRelation, Profile>[] =
       await this.relationsService.getAllFriendsRelations(req.user.userId);
 
     return pairs.map((pair: Pair<UserRelation, Profile>) => {
+      if (req.user === undefined) throw new UnauthorizedException();
+
       return {
         profile: profileToProfileResponse(pair.second),
         friends_status: relationToFriendsStatus(pair.first, req.user.userId),
@@ -98,9 +107,10 @@ export class RelationsController {
   })
   @Get('/username/:username')
   async getRelationWithOther(
-    @Req() req: any,
+    @Req() req: Express.Request,
     @Param('username') otherUsername: string,
   ): Promise<RelationResponse> {
+    if (req.user === undefined) throw new UnauthorizedException();
     const relation = await this.relationsService.getRelation(
       req.user.userId,
       otherUsername,
@@ -127,7 +137,8 @@ export class RelationsController {
   })
   @HttpCode(HttpStatus.OK)
   @Post('/block')
-  async block(@Req() req: any, @Body() addFriendDto: FriendDto) {
+  async block(@Req() req: Express.Request, @Body() addFriendDto: FriendDto) {
+    if (req.user === undefined) throw new UnauthorizedException();
     await this.relationsService.blockUser(
       req.user.userId,
       addFriendDto.friend_username,
@@ -143,7 +154,8 @@ export class RelationsController {
   })
   @HttpCode(HttpStatus.OK)
   @Post('/unblock')
-  async unblock(@Req() req: any, @Body() addFriendDto: FriendDto) {
+  async unblock(@Req() req: Express.Request, @Body() addFriendDto: FriendDto) {
+    if (req.user === undefined) throw new UnauthorizedException();
     await this.relationsService.unblockUser(
       req.user.userId,
       addFriendDto.friend_username,
